requirements.txt
numpy
tensorflow
pyyaml
